DLS-10489 Don't filter out PRR
Before it was dynamically generated in the calculator even if this function returned false and said there should be no PRR. I consider this a part of the PRR calculation logic, so removing in favour of the user-supplied value. PRR question skip is no longer possible, PRR value has to be always provided.